Update MyCart Page

+++ app/Http/Controllers/User/WishlistController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\User;
 use App\Models\Wishlist;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 
 class WishlistController extends Controller
 {
@@ -14,8 +15,7 @@ class WishlistController extends Controller
     }//end viewWishlist function
 
     public function deleteWishlist($id){
-        $wishlist = Wishlist::findOrfail($id);
-        $wishlist->delete();
+        $wishlist = Wishlist::where('user_id' , Auth::id())->where('id' , $id)->delete();
         return response()->json(['success' => 'Done delete Wishlist']);
     }//end viewWishlist function
 }
